Replaced DELEGATECALL with CALL on Bridge tests with VM
- Also moved Bridge VM tests to separate test class
